+++ lib/SettingJsonParser.js
@@ -19,7 +19,7 @@
 
 const fs = require('fs-extra');
 const path = require('path');
-const { deepMerge } = require('../../../../lib/util');
+const { deepMerge } = require('./util');
 
 class SettingJsonParser {
     constructor (wwwDir) {
